put Pal stuff in extra module
